chore: changed go.mod to have correct git link
