+++ fs/fs.go
@@ -1,4 +1,4 @@
-package main
+package fs
 
 import (
 
@@ -27,27 +27,3 @@ func (Dir) Attr(ctx context.Context, a *fuse.Attr) error {
 
 }
 
-
-func main() {
-
-	if len(os.Args) != 2 {
-		fmt.Println("Invalid usage. Use ./radFs <mntpoint>")
-		return
-	}
-
-	mount_point := os.Args[1]
-
-	//c is a fuse connection to dev/fuse
-	c, err := fuse.Mount(mount_point)
-
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	defer c.Close() //delay execution of Close
-
-	fs.Serve(c, FS{}) //starts listening for FS reqs
-	
-
-}
