+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/NorSomething/radFS
+module github.com/acmpesuecc/radFS
 
 go 1.25.7
